lua: move version related code to a separate module
This commit exports code, which is related to version and which was used
in box/lua/upgrade.lua, into a separate file and exports it to Lua API as
'internal.version'.

This is needed in order to avoid code duplication and will be used to
allow config module properly upgrading names for different versions.

NO_DOC=internal
NO_TEST=iternal
NO_CHANGELOG=internal